Refactor classes that have been renamed/repackaged
Refactor and move classes (to their respective modules) that have been renamed/repackaged in JDK 9.
