+++ src/classes/modules/java.base/jdk/internal/misc/JavaIOFileDescriptorAccess.java
@@ -15,19 +15,23 @@
  * See the License for the specific language governing permissions and 
  * limitations under the License.
  */
-package sun.misc;
+
+package jdk.internal.misc;
+
+import java.io.FileDescriptor;
 
 /**
- * <2do> this is not supported yet
+ * this is a placeholder for a Java 6 class, which we only have here to
+ * support both Java 1.5 and 6 with the same set of env/ classes
+ *
+ * see sun.msic.SharedSecrets for details
+ *
+ * <2do> THIS IS GOING AWAY AS SOON AS WE OFFICIALLY SWITCH TO JAVA 6
  */
-public interface JavaNioAccess {
-  
-  interface BufferPool {
-    String getName();
-    long getCount();
-    long getTotalCapacity();
-    long getMemoryUsed();
-  }
 
-  BufferPool getDirectBufferPool();
+public interface JavaIOFileDescriptorAccess {
+    public void set(FileDescriptor fdObj, int fd);
+    public int get(FileDescriptor fdObj);
+    public void setHandle(FileDescriptor fdObj, long h);
+    public long getHandle(FileDescriptor fdObj);
 }
