+++ src/classes/modules/java.base/jdk/internal/misc/JavaLangAccess.java
@@ -16,21 +16,36 @@
  * limitations under the License.
  */
 
-package sun.misc;
+package jdk.internal.misc;
 
-//import java.io.Console;
-import java.nio.charset.Charset;
+import sun.nio.ch.Interruptible;
+import sun.reflect.ConstantPool;
+import sun.reflect.annotation.AnnotationType;
 
 /**
  * this is a placeholder for a Java 6 class, which we only have here to
  * support both Java 1.5 and 6 with the same set of env/ classes
  *
- * see sun.msic.SharedSecrets for details
+ * see sun.misc.SharedSecrets for details
  *
  * <2do> THIS IS GOING AWAY AS SOON AS WE OFFICIALLY SWITCH TO JAVA 6
  */
-public interface JavaIOAccess {
-    //public Console console(); // not in Java 1.5, so we skip for now
-    public Runnable consoleRestoreHook();
-    public Charset charset();
+
+public interface JavaLangAccess {
+
+    ConstantPool getConstantPool(Class<?> klass);
+
+    void setAnnotationType(Class<?> klass, AnnotationType annotationType);
+
+    AnnotationType getAnnotationType(Class<?> klass);
+
+    <E extends Enum<E>> E[] getEnumConstantsShared(Class<E> klass);
+
+    void blockedOn(Thread t, Interruptible b);
+
+    void registerShutdownHook(int slot, Runnable r);
+    
+    int getStackTraceDepth(Throwable t);
+    
+    StackTraceElement getStackTraceElement(Throwable t, int i);
 }
