+++ src/classes/modules/java.base/jdk/internal/misc/Unsafe.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and 
  * limitations under the License.
  */
-package sun.misc;
+package jdk.internal.misc;
 
 import java.lang.reflect.Field;
